#68 - Renamed configure creature and fixed build errors
