+++ apps/spa-3ncount3r/module-federation.config.js
@@ -2,5 +2,6 @@ module.exports = {
   name: 'spa-3ncount3r',
   exposes: {
     './Module': './src/remote-entry.ts',
+    './menu': './src/remote-exports/encounters-menu/encounters-menu.ts'
   },
 };
