+++ apps/spa-3ncount3r/src/components/modules/search-creatures/search-creatures.tsx
@@ -5,7 +5,7 @@ import React, { useRef, useState } from 'react';
 import { useOnClickOutside } from '../../../hooks/useOnClickOutside.hook';
 import { doCreatureSearch } from '../../../services/search.service';
 import { useEncounterContext } from '../../../providers/encounterContext/encounter.context-provider';
-import { ConfigureCreatureModal } from '../../modals/configure-creature/configureCreature.modal';
+import { ConfigureCreatureModal } from '../../modals/configure-creature/configure-creature.modal';
 import SearchCreatureListItem from '../search-creature-list-item/search-creature-list-item';
 
 import { CreatureViewModel } from '../../../view-models/creature.view-model';
