+++ apps/spa-3ncount3r/src/remote-exports/app/App.tsx
@@ -1,4 +1,3 @@
-import './App.css';
 import React from 'react';
 import {
   RouterProvider
