+++ package.json
@@ -26,6 +26,7 @@
     "react-router-dom": "^6.14.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
+    "uuid4": "^2.0.3",
     "zone.js": "~0.13.0"
   },
   "devDependencies": {
@@ -59,6 +60,7 @@
     "@types/node": "18.14.2",
     "@types/react": "18.2.13",
     "@types/react-dom": "18.2.6",
+    "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "^5.58.0",
     "@typescript-eslint/parser": "^5.58.0",
     "babel-jest": "^29.4.1",
